// Webvantage.Angular/ClientApp/src/app/modules/proofing/constants/types/asset-info-table/asset-table.constants.ts
export const enum ASSET {
  TITLE = 'Title',
  STATUS = 'Status',
  DESCRIPTION = 'Description',
  VERSION = 'Version',
  CREATED_DATE = 'Created Date',
  CREATED_BY = 'Created By',
  LAST_MODIFIED = 'Last Modified',
  ASSET_URL = 'Asset URL',
  TAGS = 'Tags',

  LATEST_DOCUMENT = "Latest Document",
  LATEST_MARKUP_DATE = "Latest Markup Date",
  LATEST_MARKUP_FULL_NAME =  "Latest Markup User",
  TOTAL_APPROVES = "Total Approves",
  TOTAL_DEFERS = "Total Defers",
  TOTAL_MARKUPS =  "Total Markups",
  TOTAL_REJECTS = "Total Rejects",
  TOTAL_VERSIONS = "Total Versions",
  TOTAL_VERSIONS_FOR_ALERT_ID =  "Total Versions For Proof",
}